// src/types/hr.leave.d.ts
import { TActivityMixin, TDatabaseAuditDetails, TMailThread, OdooIdNameTuple, OdooFalseOr } from './global';

export type THrLeave = {
  id: number;
  active?: boolean;
  name: OdooFalseOr<string>;
  private_name: OdooFalseOr<string>;
  state: string;
  report_note?: boolean;
  user_id: OdooIdNameTuple;
  manager_id: OdooFalseOr<OdooIdNameTuple>;
  is_user_only_responsible?: boolean;
  holiday_status_id: OdooIdNameTuple;
  color: number;
  validation_type: string;
  employee_id: OdooIdNameTuple;
  employee_company_id: OdooIdNameTuple;
  company_id: OdooIdNameTuple;
  active_employee: boolean;
  tz_mismatch: boolean;
  tz: string;
  department_id: OdooFalseOr<OdooIdNameTuple>;
  notes: OdooFalseOr<string>;
  resource_calendar_id: OdooIdNameTuple;
  date_from: string;
  date_to: string;
  number_of_days: number;
  number_of_hours: number;
  last_several_days: boolean;
  number_of_days_display?: number;
  number_of_hours_display?: number;
  number_of_hours_text?: string;
  duration_display: string;
  meeting_id: OdooFalseOr<OdooIdNameTuple>;
  parent_id?: OdooFalseOr<OdooIdNameTuple>;
  linked_request_ids?: number[];
  holiday_type?: string;
  employee_ids?: number[];
  multi_employee?: boolean;
  category_id?: OdooFalseOr<OdooIdNameTuple>;
  mode_company_id?: OdooFalseOr<OdooIdNameTuple>;
  first_approver_id: OdooFalseOr<OdooIdNameTuple>;
  second_approver_id: OdooFalseOr<OdooIdNameTuple>;
  can_reset: boolean;
  can_approve: boolean;
  can_cancel: boolean;
  attachment_ids: number[];
  supported_attachment_ids: number[];
  supported_attachment_ids_count: number;
  all_employee_ids?: number[];
  leave_type_request_unit: string;
  leave_type_support_document: boolean;
  request_date_from: string;
  request_date_to: string;
  request_hour_from: number;
  request_hour_to: number;
  request_date_from_period: string;
  request_unit_half: boolean;
  request_unit_hours: boolean;
  is_hatched: boolean;
  is_striked: boolean;
  has_mandatory_day: boolean;
  leave_type_increases_duration: boolean;
  overtime_id: OdooFalseOr<OdooIdNameTuple>;
  employee_overtime: number;
  overtime_deductible: boolean;
  display_name?: string;
  timesheet_ids?: number[];
  payslip_state?: string;
} & TActivityMixin &
  TMailThread &
  TDatabaseAuditDetails;